pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "discord-py>=2.5.2",
    "gtts>=2.5.4",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
]
